Fix a bug where non-users could access article creation and deletion
Leftover from a previous commit.

+++ src/uts/wsd/controller/action/ArticleDeleteAction.java
@@ -15,11 +15,16 @@ public class ArticleDeleteAction implements Action {
     private HttpServletRequest request;
 
     public String execute(HttpServletRequest request, HttpServletResponse response) throws Exception {
+        this.request = request;
+
+        Author user = (Author)request.getSession().getAttribute("user");
         int id = Integer.parseInt(request.getParameter("id"));
         
         ArticleDAO articleDao = new ArticleDAOImpl(request.getSession().getServletContext());
         Article article = articleDao.findArticle(id);
-        articleDao.deleteArticle(article);
+
+        if (article.getAuthor().getId() == user.getId())
+            articleDao.deleteArticle(article);
 
         return "";
     }
